Cleaned up Ogre conversion utilities.
