+++ include/or_interactivemarker/util/ogre_conversions.h
@@ -3,7 +3,7 @@
 Copyright (c) 2015, Carnegie Mellon University
 All rights reserved.
 
-Authors: Michael Koval <mkoval@cs.cmu.edu>
+Authors: Matthew Klingensmith <mklingen@cs.cmu.edu>
 
 Redistribution and use in source and binary forms, with or without
 modification, are permitted provided that the following conditions are
@@ -29,39 +29,25 @@ THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
 OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 
 *************************************************************************/
-#ifndef ORCONVERSIONS_H_
-#define ORCONVERSIONS_H_
-#include <std_msgs/ColorRGBA.h>
+#ifndef OGRE_CONVERSIONS_H_
+#define OGRE_CONVERSIONS_H_
+
+#include <OgrePose.h>
 #include <geometry_msgs/Pose.h>
-#include <geometry_msgs/Point.h>
-#include <geometry_msgs/Quaternion.h>
-#include <geometry_msgs/Vector3.h>
-#include <openrave/openrave.h>
+#include <openrave/geometry.h>
 
 namespace or_interactivemarker {
 namespace util {
 
-extern std::string const kDefaultWorldFrameId;
+template <class Scalar>
+OpenRAVE::geometry::RaveVector<Scalar> toORVector(Ogre::Vector3 const &vec);
+template <class Scalar>
+Ogre::Vector3 toOgreVector(OpenRAVE::geometry::RaveVector<Scalar> const &vec);
 
-// OpenRAVE to ROS
 template <class Scalar>
-std_msgs::ColorRGBA toROSColor(OpenRAVE::RaveVector<Scalar> const &color);
+OpenRAVE::geometry::RaveVector<Scalar> toORQuaternion(Ogre::Quaternion const &quat);
 template <class Scalar>
-geometry_msgs::Vector3 toROSVector(OpenRAVE::RaveVector<Scalar> const &or_vector);
-template <class Scalar>
-geometry_msgs::Pose toROSPose(OpenRAVE::RaveTransform<Scalar> const &or_pose);
-template <class Scalar>
-geometry_msgs::Point toROSPoint(OpenRAVE::RaveVector<Scalar> const &or_point);
-template <class Scalar>
-geometry_msgs::Quaternion toROSQuaternion(OpenRAVE::RaveVector<Scalar> const &or_quat);
-
-// ROS to OpenRAVE
-template <class Scalar>
-OpenRAVE::RaveVector<Scalar> toORPoint(geometry_msgs::Point const &point);
-template <class Scalar>
-OpenRAVE::RaveVector<Scalar> toORQuaternion(geometry_msgs::Quaternion const &quat);
-template <class Scalar>
-OpenRAVE::RaveTransform<Scalar> toORPose(geometry_msgs::Pose const &pose);
+Ogre::Quaternion toOgreQuaternion(OpenRAVE::geometry::RaveVector<Scalar> const &vec);
 
 }
 }
