+++ src/InteractiveMarkerViewer.cpp
@@ -34,7 +34,7 @@ OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 #include <boost/algorithm/string/trim.hpp>
 #include <interactive_markers/interactive_marker_server.h>
 #include "util/ScopedConnection.h"
-#include "util/or_conversions.h"
+#include "util/ros_conversions.h"
 #include "InteractiveMarkerViewer.h"
 
 using boost::format;
