+++ src/markers/KinBodyLinkMarker.cpp
@@ -30,7 +30,7 @@ OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 
 *************************************************************************/
 #include "markers/KinBodyLinkMarker.h"
-#include "util/or_conversions.h"
+#include "util/ros_conversions.h"
 
 using interactive_markers::MenuHandler;
 using visualization_msgs::InteractiveMarkerFeedbackConstPtr;
