+++ src/markers/ManipulatorMarker.cpp
@@ -34,7 +34,7 @@ OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 #include <boost/algorithm/string/predicate.hpp>
 #include <boost/range/adaptor/map.hpp>
 #include "markers/ManipulatorMarker.h"
-#include "util/or_conversions.h"
+#include "util/ros_conversions.h"
 
 using boost::format;
 using boost::str;
