+++ src/util/ros_conversions.cpp
@@ -3,7 +3,7 @@
 Copyright (c) 2015, Carnegie Mellon University
 All rights reserved.
 
-Authors: Matthew Klingensmith <mklingen@cs.cmu.edu>
+Authors: Michael Koval <mkoval@cs.cmu.edu>
 
 Redistribution and use in source and binary forms, with or without
 modification, are permitted provided that the following conditions are
@@ -31,59 +31,121 @@ OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 *************************************************************************/
 #include "util/ros_conversions.h"
 
+using OpenRAVE::dReal;
+using OpenRAVE::RaveVector;
+using OpenRAVE::RaveTransform;
+using std_msgs::ColorRGBA;
+using geometry_msgs::Vector3;
+using geometry_msgs::Pose;
+using geometry_msgs::Point;
+using geometry_msgs::Quaternion;
+
 namespace or_interactivemarker {
 namespace util {
 
-OpenRAVE::geometry::RaveVector<double> ToRaveVector(Ogre::Vector3 const & vec)
+std::string const kDefaultWorldFrameId = "/map";
+
+/*
+ * OpenRAVE to ROS
+ */
+template <class Scalar>
+ColorRGBA toROSColor(RaveVector<Scalar> const &or_color)
 {
-    return OpenRAVE::geometry::RaveVector<double>(vec.x, vec.y, vec.z);
+    ColorRGBA color;
+    color.r = or_color.x;
+    color.g = or_color.y;
+    color.b = or_color.z;
+    color.a = or_color.w;
+    return color;
 }
 
-Ogre::Vector3 ToOgreVector(OpenRAVE::geometry::RaveVector<double> const &vec)
+template <class Scalar>
+Vector3 toROSVector(RaveVector<Scalar> const &or_vector)
 {
-    return Ogre::Vector3(vec.x, vec.y, vec.z);
+    Vector3 v;
+    v.x = or_vector.x;
+    v.y = or_vector.y;
+    v.z = or_vector.z;
+    return v;
 }
 
-OpenRAVE::geometry::RaveVector<double> ToRaveQuaternion(Ogre::Quaternion const &quat)
+template <class Scalar>
+Pose toROSPose(RaveTransform<Scalar> const &or_pose)
 {
-    return OpenRAVE::geometry::RaveVector<double>(quat.x, quat.y, quat.z, quat.w);
+    Pose pose;
+    pose.position = toROSPoint<>(or_pose.trans);
+    pose.orientation = toROSQuaternion<>(or_pose.rot);
+    return pose;
 }
 
-Ogre::Quaternion ToOgreQuaternion(OpenRAVE::geometry::RaveVector<double> const &vec)
+template <class Scalar>
+Point toROSPoint(RaveVector<Scalar> const &or_point)
 {
-    return Ogre::Quaternion(vec.x, vec.y, vec.z, vec.w);
+    Point point;
+    point.x = or_point.x;
+    point.y = or_point.y;
+    point.z = or_point.z;
+    return point;
 }
 
-geometry_msgs::Pose ToGeomMsgPose(OpenRAVE::geometry::RaveTransform<double> const &transform)
+template <class Scalar>
+Quaternion toROSQuaternion(RaveVector<Scalar> const &or_quat)
 {
-    geometry_msgs::Pose toReturn;
-    toReturn.position.x = transform.trans.x;
-    toReturn.position.y = transform.trans.y;
-    toReturn.position.z = transform.trans.z;
-    toReturn.orientation.w = transform.rot.x;
-    toReturn.orientation.x = transform.rot.y;
-    toReturn.orientation.y = transform.rot.z;
-    toReturn.orientation.z = transform.rot.w;
-    return toReturn;
+    Quaternion quaternion;
+    quaternion.w = or_quat[0];
+    quaternion.x = or_quat[1];
+    quaternion.y = or_quat[2];
+    quaternion.z = or_quat[3];
+    return quaternion;
 }
 
-
-OpenRAVE::geometry::RaveTransform<double> ToRaveTransform(geometry_msgs::Pose const &pose)
+/*
+ * ROS to OpenRAVE
+ */
+template <class Scalar>
+RaveVector<Scalar> toORPoint(Point const &point)
 {
-    OpenRAVE::geometry::RaveVector<double> const quat(
-        pose.orientation.w,
-        pose.orientation.x,
-        pose.orientation.y,
-        pose.orientation.z
-    );
-    OpenRAVE::geometry::RaveVector<double> const translation(
-        pose.position.x,
-        pose.position.y,
-        pose.position.z
-    );
-
-    return OpenRAVE::geometry::RaveTransform<double>(quat, translation);
+    return RaveVector<Scalar>(point.x, point.y, point.z);
 }
+
+template <class Scalar>
+RaveVector<Scalar> toORQuaternion(Quaternion const &quat)
+{
+    RaveVector<Scalar> or_quat;
+    or_quat[0] = quat.w;
+    or_quat[1] = quat.x;
+    or_quat[2] = quat.y;
+    or_quat[3] = quat.z;
+    return or_quat;
+}
+
+template <class Scalar>
+RaveTransform<Scalar> toORPose(Pose const &pose)
+{
+    RaveTransform<Scalar> or_transform;
+    or_transform.trans = toORPoint<Scalar>(pose.position);
+    or_transform.rot = toORQuaternion<Scalar>(pose.orientation);
+    return or_transform;
+}
+
+// Explicit instantiations.
+template ColorRGBA toROSColor<float>(RaveVector<float> const &color);
+template Vector3 toROSVector<float>(RaveVector<float> const &or_vector);
+template Pose toROSPose<float>(RaveTransform<float> const &or_pose);
+template Point toROSPoint<float>(RaveVector<float> const &or_point);
+template Quaternion toROSQuaternion<float>(RaveVector<float> const &or_point);
+template RaveVector<float> toORPoint(Point const &point);
+template RaveVector<float> toORQuaternion(Quaternion const &quat);
+template RaveTransform<float> toORPose(Pose const &pose);
+
+template ColorRGBA toROSColor<double>(RaveVector<double> const &color);
+template Vector3 toROSVector<double>(RaveVector<double> const &or_vector);
+template Pose toROSPose<double>(RaveTransform<double> const &or_pose);
+template Point toROSPoint<double>(RaveVector<double> const &or_point);
+template Quaternion toROSQuaternion<double>(RaveVector<double> const &or_point);
+template RaveVector<double> toORPoint(Point const &point);
+template RaveVector<double> toORQuaternion(Quaternion const &quat);
+template RaveTransform<double> toORPose(Pose const &pose);
 
 }
 }
